UpdaterService/requirements.txt
python-json-logger==0.1.11
pyfortified-logging==0.2.1
APScheduler==3.6.3
asyncio==3.4.3
asyncpg==0.21.0
docker==4.3.1
docker-compose==1.27.4
psycopg2-binary==2.8.6
aiofiles==0.6.0
aiohttp==3.7.3
codetiming==1.2.0

PyYAML==5.3.1
sanic==20.9.0
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.4.post1
tabulate==0.8.7
PyInquirer==1.0.3
injector==0.18.4
PyInquirer==1.0.3
uvloop==0.14.0
